Add hooks to packages/ui and reorganize
