+++ packtrain-frontend/apps/admin/src/pages/Members.tsx
@@ -7,8 +7,8 @@ import {
   TextInput,
 } from "@mantine/core";
 import { CourseMember } from "@repo/api/openapi";
-import { Loading } from "@repo/ui/Loading";
-import { TableHeader, useTableData } from "@repo/ui/table/Table";
+import { Loading } from "@repo/ui/components/Loading";
+import { TableHeader, useTableData } from "@repo/ui/components/table/Table";
 import { IconSearch } from "@tabler/icons-react";
 import { useGetMembers } from "../hooks";
