+++ packtrain-frontend/apps/instructor/src/pages/migration/MigrationsLoad.tsx
@@ -16,7 +16,7 @@ import { useForm } from "@mantine/form";
 import { getApiClient } from "@repo/api/index";
 import { MasterMigration } from "@repo/api/openapi";
 import { store$ } from "@repo/api/store";
-import { Loading } from "@repo/ui/Loading";
+import { Loading } from "@repo/ui/components/Loading";
 import { useMutation } from "@tanstack/react-query";
 import React, { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
