+++ packtrain-frontend/apps/student/src/pages/ExtensionForm.tsx
@@ -16,7 +16,7 @@ import { getApiClient } from "@repo/api/index";
 import { Assignment, LateRequest } from "@repo/api/openapi";
 import { store$ } from "@repo/api/store.js";
 import { calculateNewDueDate, formattedDate } from "@repo/ui/DateUtil";
-import { Loading } from "@repo/ui/Loading";
+import { Loading } from "@repo/ui/components/Loading";
 import { useMutation } from "@tanstack/react-query";
 import { useState } from "react";
 import { useAuth } from "react-oidc-context";
