+++ packtrain-frontend/packages/ui/src/MiddlewareLayout.tsx
@@ -1,32 +1,15 @@
-import { getApiClient } from "@repo/api/index";
-import { User } from "@repo/api/openapi";
 import { store$ } from "@repo/api/store";
-import { useQuery } from "@tanstack/react-query";
 import { useEffect } from "react";
 import { useAuth } from "react-oidc-context";
 import { Outlet, useLocation, useNavigate } from "react-router-dom";
+import { useGetUser } from "./hooks";
 
 export const MiddlewareLayout = () => {
   const navigate = useNavigate();
   const location = useLocation();
   const auth = useAuth();
   const currentPage = location.pathname;
-
-  const {
-    data: userInfo,
-    error: userError,
-    status,
-  } = useQuery<User | null>({
-    queryKey: ["getUser"],
-    queryFn: () =>
-      getApiClient()
-        .then((client) => client.get_user())
-        .then((res) => res.data)
-        .catch((err) => {
-          console.log(err);
-          return null;
-        }),
-  });
+  const { data, error, status } = useGetUser(true);
 
   useEffect(() => {
     const shouldSkip =
@@ -35,19 +18,19 @@ export const MiddlewareLayout = () => {
 
     const fetchData = async () => {
       if (
-        !userInfo &&
-        !userError &&
+        !data &&
+        !error &&
         (currentPage.includes("/admin") || currentPage.includes("/instructor"))
       ) {
         navigate("/");
         window.location.href = "/";
-      } else if (auth.isAuthenticated && userError) {
+      } else if (auth.isAuthenticated && error) {
         navigate("/disabled");
       } else if (auth.isAuthenticated && store$.id.get() === undefined) {
         navigate("/select");
       } else if (
-        userInfo !== undefined &&
-        !userInfo?.admin &&
+        data !== undefined &&
+        !data?.admin &&
         currentPage.includes("/admin")
       ) {
         window.location.href = "/";
@@ -66,7 +49,7 @@ export const MiddlewareLayout = () => {
     ) {
       fetchData();
     }
-  }, [auth, userInfo, userError, navigate, currentPage, status]);
+  }, [auth, data, error, navigate, currentPage, status]);
 
   return <Outlet />;
 };
