+++ packtrain-frontend/packages/ui/src/components/footer/Footer.module.scss
