+++ packtrain-frontend/packages/ui/src/components/navbar/Navbar.module.scss
